Build: Fix WebAssembly compilation and upgrade WebGPU bindings
- Downgraded and then upgraded emsdk to ensure clean cache.

- Re-enabled EARTHCALL_WEBGPU and IMGUI_IMPL_WEBGPU_BACKEND_DAWN in CMakeLists.

- Removed ImGui #error block in imgui_impl_wgpu.cpp to unlock Dawn WebGPU bindings on Emscripten.

- Created WebGpuContext_wasm.cpp to correctly bridge Emscripten's HTML5 Canvas WebGPU initialization with Earthcall's WindowContext.

- Excluded legacy OpenGL rendering (OpenGLRenderer.cpp) and native macOS (.mm) webviews from the WASM target build.

- Safely stubbed OpenSSL, networking (ASIO), and File IO subsystems with __EMSCRIPTEN__ guards.

+++ src/Identity/SingularId.hpp
@@ -104,7 +104,7 @@ template <>
 struct hash<Identity::SingularId> {
     size_t operator()(const Identity::SingularId& id) const noexcept {
         // FNV-1a over the raw bytes; ids are already uniformly random.
-        size_t h = 1469598103934665603ull;
+        uint64_t h = 1469598103934665603ull;
         for (uint8_t b : id.bytes()) {
             h ^= b;
             h *= 1099511628211ull;
